+++ src/entry.rs
@@ -138,14 +138,14 @@ mod tests {
     impl EzApp for Marker {
         type Request = String;
         type Response = String;
+        type ReadRequest = ();
+        type ReadResponse = ();
 
         async fn apply(&mut self, req: String) -> String {
             req
         }
 
-        fn read(&self, _key: &str) -> Option<serde_json::Value> {
-            None
-        }
+        fn read(&self, _req: ()) {}
     }
 
     #[test]